Move `lfilter::check_and_get_axis*` to `arraytools::`
These are fairly generic functions that parse user input into valid
values of usizes (in their respective context being the arrays they act
on), and thus are more suited to be in array tools.
